Backend/requirements.txt
google-api-python-client==2.148.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
msal==1.31.0
requests==2.32.3
python-dotenv==1.0.1
pymongo[srv]
pydantic==1.10.11
dnspython